Fixed a whole bunch of bugs

+++ frontend/src/admin_components/FileSelector.js
@@ -1,17 +1,18 @@
 import { useState } from "react";
-import { Button, Container } from "react-bootstrap";
+import { Button, Container, Modal } from "react-bootstrap";
 
 import styles from "./FileSelector.module.scss"
 
 function FileSelector({ creds, type, onSelect, artist, data, index }) {
 
   const [artists, setArtists] = useState(data);
+  const [showImageSelect, setShowImageSelect] = useState(false);
 
   function handleUpload(e) {
     const files = e.target.files;
     var NameArray = [];
 
-    for (let i = 0; i < files.length; i++){
+    for (let i = 0; i < files.length; i++) {
       NameArray.push(Date.now().toString() + "_" + files[i].name);
     }
 
@@ -39,13 +40,13 @@ function FileSelector({ creds, type, onSelect, artist, data, index }) {
       })
   }
 
-  function handleAddSlide(NameArray){
-    for(var i = 0; i < NameArray.length; i++){
+  function handleAddSlide(NameArray) {
+    for (var i = 0; i < NameArray.length; i++) {
       data[index].images.push(NameArray[i]);
     }
 
     setArtists([...data]);
-   
+
     const url = process.env.REACT_APP_BACKEND + "admin/update/artists";
     let headers = new Headers();
     headers.append(
@@ -68,7 +69,7 @@ function FileSelector({ creds, type, onSelect, artist, data, index }) {
   }
 
   function handleDelete(src, idx) {
-    let url = process.env.REACT_APP_BACKEND + "admin/delete_image/" + type+ "/" +src;
+    let url = process.env.REACT_APP_BACKEND + "admin/delete_image/" + type + "/" + src;
 
     let headers = new Headers();
 
@@ -82,17 +83,17 @@ function FileSelector({ creds, type, onSelect, artist, data, index }) {
       headers
     }).catch(err => alert(err))
       .then(data => {
-        if(type === "artist"){
+        if (type === "artist") {
           handleDeleteSlide(idx);
         }
       })
   }
 
-  function handleDeleteSlide(idx){
+  function handleDeleteSlide(idx) {
 
     data[index].images.splice(idx, 1);
     setArtists([...data]);
-   
+
     const url = process.env.REACT_APP_BACKEND + "admin/update/artists";
     let headers = new Headers();
     headers.append(
@@ -137,21 +138,22 @@ function FileSelector({ creds, type, onSelect, artist, data, index }) {
     }).catch(err => alert(err))
       .then(data => {
         handleDelete(artist.imageUrl, 0);
-        if(index === 999){
+        if (index === 999) {
           handleUpdateTeam(filename);
-        }else{
+        } else {
           handleUpdateWorker(filename);
         }
         e.target.value = "";
+      setArtists({...data});
       })
   }
 
-  function handleUpdateWorker(filename){
+  function handleUpdateWorker(filename) {
     var toBeDeleted = data.members[index].imageUrl;
-    data.members[index].imageUrl = filename;
+    artist.imageUrl = filename;
 
     setArtists({...data});
-   
+
     const url = process.env.REACT_APP_BACKEND + "admin/update/about";
 
     let headers = new Headers();
@@ -168,19 +170,21 @@ function FileSelector({ creds, type, onSelect, artist, data, index }) {
     fetch(url, options)
       .then((data) => {
         console.log(data);
-        handleDelete(toBeDeleted);
+        if (toBeDeleted !== "") {
+          handleDelete(toBeDeleted);
+        }
       })
       .catch((err) => {
         console.warn(err);
       });
   }
 
-  function handleUpdateTeam(filename){
+  function handleUpdateTeam(filename) {
     var toBeDeleted = data.imageUrl;
     data.imageUrl = filename;
 
-    setArtists({...data});
-   
+    setArtists({ ...data });
+
     const url = process.env.REACT_APP_BACKEND + "admin/update/about";
 
     let headers = new Headers();
@@ -205,40 +209,89 @@ function FileSelector({ creds, type, onSelect, artist, data, index }) {
   }
 
 
-  if(type === "artist"){
+  if (type === "artist") {
     return (
-      <Container className={styles.container}>  
-        {artist.images ? 
+      <Container className={styles.container}>
+        {artist.images ?
           artist.images.map((src, idx) => {
             return (
               <div key={src} className={styles.imageContainer}>
-                <img onClick={_=>onSelect(src)} alt={src} className={styles.image} width={100} height={100} src={process.env.REACT_APP_BACKEND + "images/" + type + "/" + src} />
-                <Button onClick={_=>handleDelete(src, idx)} className={styles.deleteButton} variant="danger">-</Button>
+                <img onClick={_ => onSelect(src)} alt={src} className={styles.image} width={100} height={100} src={process.env.REACT_APP_BACKEND + "images/" + type + "/" + src} />
+                <Button onClick={_ => handleDelete(src, idx)} className={styles.deleteButton} variant="danger">-</Button>
                 <span>{src}</span>
               </div>
             )
           })
-          :""
+          : ""
         }
         <div className={styles.imageContainer + " " + styles.addBtn}>
           <input type="file" onChange={e => handleUpload(e)} />
         </div>
       </Container>
-    ); 
+    );
+  } else if(index !== 999){
+    return (
+      <>
+        <Button
+          variant="info"
+          onClick={(_) => setShowImageSelect(true)}
+          key={artist + index + "Button"}
+        >
+          Change
+                    </Button>
+        <Modal key={artist + index} show={showImageSelect} onHide={() => setShowImageSelect(false)}>
+          <Modal.Header
+            closeButton
+            onClick={(_) => setShowImageSelect(false)}
+          >
+            <Modal.Title>Choose an image</Modal.Title>
+          </Modal.Header>
+          <Modal.Body>
+            <Container className={styles.container}>
+              {artist.imageUrl ?
+                <div key={artist.imageUrl} className={styles.imageContainer}>
+                  <img onClick={_ => onSelect(artist.imageUrl)} alt={artist.imageUrl} className={styles.image} width={100} height={100} src={process.env.REACT_APP_BACKEND + "images/" + type + "/" + artist.imageUrl} />
+                  <span>{artist.imageUrl}</span>
+                </div>
+                : ""
+              }
+              <div className={styles.imageContainer + " " + styles.addBtn}>
+                <input type="file" onChange={e => handleUpdate(e)} />
+              </div>
+            </Container>
+          </Modal.Body>
+          <Modal.Footer>
+            <p>Click on an image to select</p>
+            {/* <Button
+                          variant="secondary"
+                          onClick={(_) => setShowImageSelect(false)}
+                        >
+                          Cancel
+                        </Button>
+                        <Button
+                          variant="primary"
+                          onClick={(_) => setShowImageSelect(false)}
+                        >
+                          Save Changes
+                        </Button> */}
+          </Modal.Footer>
+        </Modal>
+      </>
+    );
   }else{
     return(
-    <Container className={styles.container}>     
-      {artist.imageUrl ?     
-        <div key={artist.imageUrl} className={styles.imageContainer}>
-          <img onClick={_=>onSelect(artist.imageUrl)} alt={artist.imageUrl} className={styles.image} width={100} height={100} src={process.env.REACT_APP_BACKEND + "images/" + type + "/" + artist.imageUrl} />
-          <span>{artist.imageUrl}</span>
+      <Container className={styles.container}>
+        {artist.imageUrl ?
+          <div key={artist.imageUrl} className={styles.imageContainer}>
+            <img onClick={_ => onSelect(artist.imageUrl)} alt={artist.imageUrl} className={styles.image} width={100} height={100} src={process.env.REACT_APP_BACKEND + "images/" + type + "/" + artist.imageUrl} />
+            <span>{artist.imageUrl}</span>
+          </div>
+          : ""
+        }
+        <div className={styles.imageContainer + " " + styles.addBtn}>
+          <input type="file" onChange={e => handleUpdate(e)} />
         </div>
-        :"" 
-      }
-      <div className={styles.imageContainer + " " + styles.addBtn}>
-        <input type="file" onChange={e => handleUpdate(e)} />
-      </div>
-    </Container>
+      </Container>
     );
   }
 }
